requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
argparse
python-dotenv